test(E2E test):  E2Eテストファイルを集約
